Make I2C polling dynamic based on update freq. Add error check on bytes avialable.
This allows us to more agressively check for I2C data as the user increases the update rate.

Also added crazy too big error check on bytes available. On the Artemis, there is sometimes the first bit of the bytes available erroneously set to 1. This causes a very large amount of bytes to be read and many valid packets to not be acted upon.

There are still some dropped packets but getting cleaner at 10Hz.