Adding all the keys... See description for full details.
This commit adds _all_ the keys from the ZED-F9P Interface Description. The ones Morten extracted plus all the others. They're included as const uint32_t. I was puzzled about why there were no duplications with the existing keys. And the reason is that the existing keys are all prefixed with UBLOX_. So, we have a decision to make. Leave them as-is? Or add UBLOX_ to all of them? Or remove UBLOX_ from all of them?